feat(game-tracker): full implementation for `init` and `commit`
**Why**

Enable the ability to create games and commit actions, which
will allow us to leverage git to stack, revert and rewrite
history. Functionality essential to games where players can react
to opponents actions, but are not able to do so in real-time.

Existing code was borderline pseudo-code implementation, only
intended as a placeholder.

Allow the game tracker to actually create a game directory and
initialize a git repo, and to commit changes to the repo.

**What**

 - Create new directory for game when initialized

 - Create new git repository in that directory, including
   file with basic metadata.

 - Implement ability to commit changes

 - Add ability to permanently delete the game from the filesystem

 - Add test coverage